+++ collective/dancing/browser/channel.py
@@ -40,6 +40,7 @@ from collective.dancing.composer import check_email
 from collective.dancing.browser import controlpanel
 from collective.dancing.browser.interfaces import ISendAndPreviewForm
 from collective.dancing.utils import switch_on
+from collective.dancing.interfaces import ISubscriptionsFromScriptChannel
 
 
 def simpleitem_wrap(klass, name):
@@ -399,28 +400,72 @@ def parseSubscriberCSVFile(subscriberdata, composer,
         reader = csv.reader(data, delimiter=str(delimiter))
         subscriberslist = []
         errorcandidates = []
+        sections = []
         for index, parsedline in enumerate(reader):
             if index == 0:
                 if header_row_present:
                     fields = parsedline
-                    continue
                 else:
                     fields = field.Fields(composer.schema).keys()
-            if len(parsedline)<len(fields):
+                    # could be import old csv that does not have sections column
+                    if len(parsedline) > len(fields):
+                        section_length = len(parsedline) - len(fields)
+                        for i in range(section_length):
+                            fields.append('section')
+                # combine all the sections into one section list
+                section_row_fields = []
+                section_index = -1
+                for i in range(len(fields)):
+                    if fields[i] != 'section':
+                        section_row_fields.append(fields[i])
+                    elif section_index < 0:
+                        section_index = i
+                        section_row_fields.append(fields[i])
+                if header_row_present:
+                    continue
+
+            if len(parsedline) < len(fields):
                 pass
             else:
                 try:
-                    subscriber = dict(zip(fields,\
-                       map(lambda x:x.decode(charset), parsedline)))
+                    # combine all the sections into one section list
+                    new_fields = fields
+                    old_parsedline = parsedline
+                    parsedline = []
+                    section_field = []
+                    for i in range(len(fields)):
+                        if fields[i] != 'section':
+                            parsedline.append(old_parsedline[i])
+                        else:
+                            section_field.append(
+                                old_parsedline[i].decode(charset))
+                    if section_index > -1:
+                        new_fields = section_row_fields
+                        parsedline.insert(section_index, section_field)
+
+                    subscriber = dict(
+                        zip(new_fields, map(lambda x: x.decode(charset)
+                            if isinstance(x, basestring)
+                            else x, parsedline)))
+
+                    # splits 'sections' data from subscriber
+                    if 'section' in subscriber:
+                        section = dict([
+                            ('email', subscriber['email']),
+                            ('section', subscriber['section'])])
+                        sections.append(section)
+                        del subscriber['section']
+
                     subscriber['email'] = subscriber['email'].lower()
                     check_email(subscriber['email'])
                 except:
                     errorcandidates.append(subscriber['email'])
                 else:
                     subscriberslist.append(subscriber)
-        return subscriberslist, errorcandidates
+        return subscriberslist, errorcandidates, sections
     except Exception, e:
         return _(u"Error importing subscriber file. %s" % e), []
+
 
 class ExportCSV(BrowserView):
 
@@ -430,17 +475,44 @@ class ExportCSV(BrowserView):
                                                          'utf-8').upper()
         self.request.response.setHeader('Content-Type',
                                         'text/csv; charset=%s' % charset)
-        self.request.response.setHeader('Content-disposition',
+        self.request.response.setHeader(
+            'Content-disposition',
             'attachment; filename=subscribers_%s_%s.csv' % (self.context.id,
-                                      datetime.date.today().strftime("%Y%m%d")))
+            datetime.date.today().strftime("%Y%m%d")))
         res = cStringIO.StringIO()
         writer = csv.writer(res, dialect=csv.excel, delimiter=csv_delimiter)
+        collectors_keys = []
+        if self.context.collector:
+            selected_collector = self.context.collector.id
+            if selected_collector in self.context.collectors.keys():
+                for optional in \
+                        self.context.collectors[selected_collector] \
+                        .get_optional_collectors():
+                    title_id = optional.title.strip() + ' (' + optional.id + ')'
+                    collectors_keys.append(title_id)
         for format in self.context.composers.keys():
-            for subscription in tuple(self.context.subscriptions.query(format=format)):
+            composers_keys = field.Fields(
+                self.context.composers[format].schema).keys()
+            # add csv header row
+            writer.writerow(composers_keys + ['section'] * len(collectors_keys))
+            for subscription in tuple(
+                    self.context.subscriptions.query(format=format)):
                 row = []
-                for item in field.Fields(self.context.composers[format].schema).keys():
+                for item in composers_keys:
                     v = subscription.composer_data.get(item) or ''
                     row.append(self._convertValue(v, charset))
+                selected_collectors = {}
+                if 'selected_collectors' in subscription.collector_data:
+                    selected_collectors = dict([
+                        (c.title.strip() + ' (' + c.id + ')', c.title.strip())
+                        for c in subscription.collector_data[
+                            'selected_collectors']])
+                for item in collectors_keys:
+                    if item in selected_collectors:
+                        # the csv section field format is "title (id)"
+                        row.append(self._convertValue(item, charset))
+                    else:
+                        row.append('')
 
                 writer.writerow(row)
         return res.getvalue()
@@ -547,12 +619,51 @@ class UploadForm(crud.AddForm):
         delimiter = data.get('csv_delimiter', csv_delimiter)
         if subscriberdata is None:
             return _(u"File was not given.")
-        subscribers, errorcandidates = parseSubscriberCSVFile(subscriberdata,
-                                self.context.composer,
-                                header_row_present=header_row_present,
-                                delimiter=delimiter)
+        subscribers, errorcandidates, sections = parseSubscriberCSVFile(
+            subscriberdata,
+            self.context.composer,
+            header_row_present=header_row_present,
+            delimiter=delimiter)
         if not type(subscribers)==type([]):
             return _(u"File has incorrect format.")
+
+        # generate new selected collectors
+        new_sections = {}
+        if self.mychannel.collector and \
+                self.mychannel.collector.id in self.mychannel.collectors.keys():
+            optional_collectors = self.mychannel.collectors[
+                self.mychannel.collector.id].get_optional_collectors()
+            collectors_dict_title = {}
+            collectors_dict_id = {}
+            for optional_collector in optional_collectors:
+                collectors_dict_title[optional_collector.title] = \
+                    optional_collector
+                collectors_dict_id[optional_collector.id] = optional_collector
+
+            for section in sections:
+                if 'section' in section and 'email' in section:
+                    sections_data = section['section']
+                    new_section_data = []
+                    for section_data in sections_data:
+                        if not section_data:
+                            continue
+                        # section data format is "title (id)" or "title"
+                        if section_data.split(' ')[-1][0] == '(' and \
+                                section_data.split(' ')[-1][-1] == ')':
+                            # use id
+                            # check if title (id)
+                            section_id = section_data.split(' ')[-1][1:-1]
+                            if section_id in collectors_dict_id:
+                                new_section_data.append(
+                                    collectors_dict_id[section_id])
+                        else:
+                            # use title
+                            # section data format is "title"
+                            if section_data in collectors_dict_title:
+                                new_section_data.append(
+                                    collectors_dict_title[section_data])
+                    new_sections[section['email']] = sets.Set(new_section_data)
+
         added = 0
         new_and_updated = []
         notadded = len(errorcandidates)
@@ -580,6 +691,17 @@ class UploadForm(crud.AddForm):
                         if 'selected_collectors' in old_collector_data \
                            and old_collector_data['selected_collectors']:
                             item.collector_data = old_collector_data
+                    # add new section selection
+                    if subscriber_data['email'] in new_sections:
+                        new_selected_collectors = \
+                            new_sections[subscriber_data['email']]
+                        if len(new_selected_collectors):
+                            item.collector_data = dict([(
+                                'selected_collectors',
+                                new_selected_collectors)])
+                        else:
+                            item.collector_data = {}
+
                     added += 1
                 except Exception, e: # XXX refine which exceptions we want to catch
                     # TODO; put some information about error e into the message
@@ -766,3 +888,18 @@ class ManageChannelView(BrowserView):
         return wrapper % \
                ("\n".join((template % (id(msg), zope.i18n.translate(msg, context=self.request), id(msg), html)
                            for (msg, html) in fieldsets)))
+
+
+class EditScriptChannelForm(EditChannelForm):
+
+    @property
+    def fields(self):
+
+        fields = super(EditScriptChannelForm, self).fields
+        fields += z3c.form.field.Fields(
+            ISubscriptionsFromScriptChannel).select('script_path')
+        return fields
+
+
+class ManageScriptChannelView(ManageChannelView):
+    edit_form = EditScriptChannelForm